migrated events/guilds controllers to typescript
