+++ packages/server/controllers/events.ts
@@ -1,16 +1,17 @@
 // const postgres = require("../utils/postgres");
-const prisma = require("../utils/prisma");
+import prisma from "../utils/prisma";
+import { Event } from ".prisma/client/index";
 
-async function getAllEvents() {
+async function getAllEvents(): Promise<Event[]> {
   return prisma.event.findMany();
 }
 
-async function getEvents(limit, action, eventId) {
-  let events;
+async function getEvents(limit: number, action: string, eventId: string): Promise<Event[]> {
+  let events: Event[];
 
   switch (action) {
     case "prev":
-      events = await prisma.events.findMany({
+      events = await prisma.event.findMany({
         take: -1 * limit,
         skip: 1, // skip cursor (TODO: check if skip 1 is needed for previous page queries)
         cursor: {
@@ -22,7 +23,7 @@ async function getEvents(limit, action, eventId) {
       });
       break;
     case "next":
-      events = await prisma.events.findMany({
+      events = await prisma.event.findMany({
         take: limit,
         skip: 1, // skip cursor
         cursor: {
@@ -35,7 +36,7 @@ async function getEvents(limit, action, eventId) {
       break;
     // first request made for first page, no action in cursor present
     default:
-      events = await prisma.events.findMany({
+      events = await prisma.event.findMany({
         take: limit,
         orderBy: {
           eventId: "asc",
@@ -47,14 +48,14 @@ async function getEvents(limit, action, eventId) {
   return events;
 }
 
-async function getPages(limit) {
-  const totalEvents = await prisma.events.count();
+async function getPages(limit: number): Promise<number> {
+  const totalEvents = await prisma.event.count();
   const totalPages = Math.ceil(totalEvents / limit);
   return totalPages;
 }
 
-async function getEvent(eventId) {
-  const event = await prisma.events.findUnique({
+async function getEvent(eventId: string): Promise<Event | null> {
+  const event = await prisma.event.findUnique({
     where: {
       eventId: eventId,
     },
