--- packages/server/controllers/guilds.js
@@ -1,20 +1,0 @@
-const prisma = require("../prisma/prisma");
-
-async function getAllGuilds() {
-  const query = await prisma.guilds.findMany();
-  return query;
-}
-
-async function getGuild(guildId) {
-  const query = await prisma.guilds.findFirst({
-    where: {
-      guildId: guildId,
-    },
-  });
-  return query;
-}
-
-module.exports = {
-  getGuild,
-  getAllGuilds,
-};
